+++ src/base/TreeUtils.ts
@@ -14,6 +14,13 @@ class TreeUtils {
       .pop()
   }
 
+  // todo: refactor so instead of str input takes an array of cells(strings) and scans each indepndently.
+  static _chooseDelimiter(str: string) {
+    const del = " ,|\t;^%$!#@~*&+-=_:?.{}[]()<>/".split("").find(idea => !str.includes(idea))
+    if (!del) throw new Error("Could not find a delimiter")
+    return del
+  }
+
   static flatten(arr: any) {
     if (arr.flat) return arr.flat()
     return arr.reduce((acc: any, val: any) => acc.concat(val), [])
@@ -212,6 +219,40 @@ class TreeUtils {
       lines--
     }
     return str
+  }
+
+  // adapted from https://gist.github.com/blixt/f17b47c62508be59987b
+  // 1993 Park-Miller LCG
+  static _getPRNG(seed: number) {
+    return function() {
+      seed = Math.imul(48271, seed) | 0 % 2147483647
+      return (seed & 2147483647) / 2147483648
+    }
+  }
+
+  private static _tickTime: number
+  // todo: clean up verbose/console log
+  static _tick(msg: string, verbose = true) {
+    if (this._tickTime === undefined) this._tickTime = Date.now() - 1000 * process.uptime()
+    const elapsed = Date.now() - this._tickTime
+    if (verbose) console.log(`${elapsed}ms ${msg}`)
+    this._tickTime = Date.now()
+    return elapsed
+  }
+
+  static _sampleWithoutReplacement(population: any[], quantity: number, seed: number) {
+    const prng = this._getPRNG(seed)
+    const sampled: { [index: number]: boolean } = {}
+    const populationSize = population.length
+    const picked = []
+    if (quantity >= populationSize) quantity = populationSize
+    while (picked.length < quantity) {
+      const index = Math.floor(prng() * populationSize)
+      if (sampled[index]) continue
+      sampled[index] = true
+      picked.push(population[index])
+    }
+    return picked
   }
 
   static arrayToMap(arr: Array<any>) {